src/insight-engine.ts
import { ChatGPTInsightsReport, ChatGTPConversations } from "./models";

export class InsightEngine {
  private conversations: ChatGTPConversations;

  constructor(conversations: ChatGTPConversations) {
    this.conversations = conversations;
  }

  private calculateMessageStats() {
    // TODO: Calculate message counts, averages, etc.
  }

  private analyzeTimePatterns() {
    // TODO: Analyze most active times
  }

  private categorizeConversationLengths() {
    // TODO: Categorize conversations by length
  }

  private analyzeModelUsage() {
    // TODO: Track which models were used
  }

  private extractTopics() {
    // TODO: Extract common topics using NLP
  }

  private analyzeSentiment() {
    // TODO: Perform sentiment analysis
  }

  public generateReport(): ChatGPTInsightsReport {
    return {
      user_analysis: {
        overview: {
          total_chats: this.conversations.length,
          total_words_used: 0,
          total_prompts: 0,
          average_prompt_length: 0,
          average_response_length: 0,
          most_active_days: [],
          most_active_hours: [],
        },
        topics_analysis: {
          most_frequent_topics: [],
          topic_histogram: {},
        },
        prompting_style: {
          common_phrases: [],
          average_prompt_tone: "Neutral",
          most_used_keywords: [],
          query_complexity_distribution: { simple: 0, moderate: 0, complex: 0 },
          prompt_efficiency: {
            average_follow_up_prompts: 0,
            percentage_prompts_needing_rephrasing: 0,
          },
        },
        sentiment_analysis: {
          overall_sentiment: "Neutral",
          sentiment_distribution: { positive: 0, neutral: 0, negative: 0 },
          common_positive_topics: [],
          common_negative_topics: [],
        },
        response_quality: {
          average_response_relevance: 0,
          common_feedback_keywords: [],
          most_helpful_topics: [],
          least_helpful_topics: [],
        },
        interaction_patterns: {
          usage_frequency: {
            daily_average: 0,
            weekly_average: 0,
            monthly_average: 0,
          },
          time_of_day_preference: {
            morning: 0,
            afternoon: 0,
            evening: 0,
            night: 0,
          },
          conversation_length_distribution: { short: 0, medium: 0, long: 0 },
        },
        advanced_insights: {
          emerging_interest: {
            topic: "",
            increase_in_prompts: 0,
            time_period: "",
          },
          redundant_queries: 0,
          most_improved_topic_knowledge: "",
          topic_diversity_score: 0,
        },
        communication_effectiveness: {
          response_clarity_score: 0,
          length_matching: { short_prompts: 0, long_prompts: 0 },
        },
        ai_utilization_patterns: {
          task_based_analysis: {
            problem_solving: 0,
            creative_exploration: 0,
            knowledge_inquiry: 0,
          },
          dependency_index: 0,
        },
        personalization_metrics: {
          tone_alignment: { preferred_tone: "", alignment_percentage: 0 },
        },
        diversity_metrics: {
          language_usage: {
            primary_language: "English",
            other_languages: [],
            percentage_non_primary_language: 0,
          },
          cross_disciplinary_interests: {
            science: 0,
            art: 0,
            business: 0,
            miscellaneous: 0,
          },
        },
        errors_and_ambiguities: {
          misunderstood_queries: 0,
          percentage_prompts_with_errors: 0,
          common_error_causes: [],
        },
      },
      visualizations: {
        topic_histogram_url: "",
        usage_timeline_url: "",
        dynamic_word_cloud_url: "",
        response_time_chart_url: "",
      },
      recommendations: {
        tips_to_improve: [],
        suggested_topics_to_explore: [],
        recommended_interaction_changes: [],
        experimentation_suggestions: [],
      },
    };
  }
}
